Feature/improve ui

+++ ui/project_item.py
@@ -3,9 +3,17 @@
 This script contains ProjectItem.
 """
 
-from PyQt5.QtWidgets import QMenu, QAction, QTreeWidgetItem, QPushButton
+from os.path import dirname, samefile
+
+from ..backend.qgis_server import QgisServer
+from ..backend.exceptions import Unauthorized
+
+from PyQt5.QtCore import QSettings
+from PyQt5.QtWidgets import QMenu, QAction, QTreeWidgetItem, QMessageBox,\
+                            QInputDialog
 
 from qgis.core import QgsProject
+from qgis.gui import QgsMessageBar
 
 from ..utils import safe_close, str2int
 
@@ -35,12 +43,6 @@ class ProjectItem(QTreeWidgetItem):
         server_item.addChild(self)
         self.setText(0, self.name)
 
-        self.publish = QPushButton('Publish')
-        server_item.treeWidget().setItemWidget(self, 1, self.publish)
-        self.publish.clicked.connect(
-            lambda: self._publish_popup()
-            )
-
     def open(self):
         """
         Open the project that this item represents in the editor.
@@ -51,6 +53,7 @@ class ProjectItem(QTreeWidgetItem):
             project.read(self.path)
             project.readProject.connect(close_project)
             project.projectSaved.connect(save_project)
+            self.clean_recent_projects()
 
         def save_project():
             self.id = self.qgis_server.upload_project(
@@ -58,6 +61,7 @@ class ProjectItem(QTreeWidgetItem):
                 self.name,
                 self.path,
             )
+            self.clean_recent_projects()
 
         def close_project():
             project = QgsProject.instance()
@@ -74,13 +78,34 @@ class ProjectItem(QTreeWidgetItem):
 
         def open_():
             self.open()
-        open_action = QAction('&Open project')
+        open_action = QAction('Open project')
         menu.addAction(open_action)
         open_action.triggered.connect(open_)
 
+        def rename():
+            self._rename_popup()
+        rename_action = QAction('Rename Project')
+        menu.addAction(rename_action)
+        rename_action.triggered.connect(rename)
+
+        def publish():
+            self._publish_popup()
+        publish_action = QAction('Publish project')
+        menu.addAction(publish_action)
+        publish_action.triggered.connect(publish)
+
+        menu.addSeparator()
+
         def delete():
-            self.qgis_server.delete_project(self.id)
-            self.server_item.removeChild(self)
+            confirm_dialog = QMessageBox(
+                QMessageBox.Question,
+                "Confirm project delete",
+                "Do you really want to delete this project?",
+                QMessageBox.Yes | QMessageBox.No,
+            )
+            if confirm_dialog.exec_() == QMessageBox.Yes:
+                self.qgis_server.delete_project(self.id)
+                self.server_item.removeChild(self)
         delete_action = QAction('Delete project')
         menu.addAction(delete_action)
         delete_action.triggered.connect(delete)
@@ -96,3 +121,39 @@ class ProjectItem(QTreeWidgetItem):
             )
             published['category'] = str2int(published['category'])
             self.published = published
+
+    def _rename_popup(self):
+        result = QInputDialog.getText(
+            None,
+            'Rename',
+            'Set the new name for the project '+self.name+':')
+
+        if result[1]:
+            name = result[0]
+            try:
+                result = self.qgis_server.upload_project(
+                    self.id,
+                    name,
+                    path=None
+                )
+                if result == self.id:
+                    self.name = name
+                    self.setText(0, self.name)
+            except Unauthorized:
+                self.iface.messageBar().pushMessage(
+                    "Error",
+                    "Couldn't rename this project",
+                    QgsMessageBar.ERROR
+                )
+
+    def clean_recent_projects(self):
+        qgis_settings = QSettings()
+        qgis_settings.beginGroup('UI/recentProjects')
+        for group in qgis_settings.childGroups():
+            if samefile(
+                    dirname(qgis_settings.value(group+'/path')),
+                    QgisServer.WRITE_DIR
+            ):
+                qgis_settings.remove(group)
+
+        qgis_settings.endGroup()
